requirements.txt
openai
pyyaml
rich
kubernetes
